+++ packages/core/src/ceramic.ts
@@ -542,7 +542,7 @@ export class Ceramic implements CeramicApi {
       indexingConfig: config.indexing,
       networkOptions,
       loadOptsOverride,
-      sync: config.indexing.enableHistoricalSync,
+      sync: config.indexing?.enableHistoricalSync,
     }
 
     const modules: CeramicModules = {