+++ packages/pinning-crust-backend/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.15.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/pinning-crust-backend@2.15.0-rc.0...@ceramicnetwork/pinning-crust-backend@2.15.0) (2023-02-28)
+
+**Note:** Version bump only for package @ceramicnetwork/pinning-crust-backend
+
+
+
+
+
 # [2.15.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/pinning-crust-backend@2.14.0-rc.0...@ceramicnetwork/pinning-crust-backend@2.15.0-rc.0) (2023-02-27)
 
 **Note:** Version bump only for package @ceramicnetwork/pinning-crust-backend